# specs/003-app-gestao-processos-aprovacao/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Gestão de Processos Condominiais API
  description: API para gestão de processos condominiais com workflow de aprovação
  version: 1.0.0
  contact:
    name: API Support
    email: support@condominio.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.condominio.com/api/v1
    description: Production server
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        user:
          $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [aprovador, visualizador, editor]
...
  responses:
...
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
...